Fix bug in ADC implementation. Add test.
This fixes the bug where an unset 'files:' entry would cause the plugin to fail.
A workaround for the bug is to set the 'files:' value to the empty string.

This patch corrects this bug and adds the test that should have existed anyway
to check the correctness of the impementation.